remove results bucket to prevent accidental data loss
Note: for future reference, accidental erasure of the bucket would require `force_destroy` to be explicitly set `true` or an AWS defect that effectively enables `force_destroy`. Simultaneously, a user would have to apply a plan that destroys the S3 bucket, such as a bucket name change or a "destroy" plan.